Added a BiometricCore
This is what I _really_ wanted to do all along :P

This core uses Veridis Biometric SDK to perform capture and matching of fingerprints.

The scanners stay off most of the time, but whenever we get a register / authenticate / wink command, they are turned on for a few seconds.
Whenever a fingerprint is captured, we set the "user Present" flag to true.

When registering, the captured fingerprint is associated with the new handle.

When authenticating, the the captured fingerprint is compared to the one associated with the handle, and if they don't match, we say "User Not Present".

If U2F messages stop for more than 5s we consider that as a "give up" and turn off the fingerprint scanners.

Unfortunately, the Veridis Biometric SDK is closed-source and slightly abandoned, so it's likely that nobody else will be able to use this U2F core :(